+++ euler579c/cube.h
@@ -8,13 +8,13 @@ using namespace std;
 class cube
 {
 public:
-	unsigned long width, height, depth;
+	long long width, height, depth;
 	set<vector3d> uvn;
 	set<vertex> vertices;
 	cube(const vector3d& U, const vector3d& V, const vector3d& N);
 
 private:
-	void add_vertex(vector<vertex>& vertices, long& minx, long& miny, long& minz, const vertex& vertex);
+	void add_vertex(vector<vertex>& vertices, long long& minx, long long& miny, long long& minz, const vertex& vertex);
 };
 
 bool operator <(const cube& lhs, const cube& rhs);
