+++ euler579c/macros.h
@@ -3,6 +3,6 @@
 #ifndef MY_MACROS
 #define MY_MACROS
 
-#define my_round(dbl) ((unsigned long)(dbl+0.5))
+#define my_round(dbl) ((long long)(dbl+0.5))
 
 #endif
