+++ euler579c/util.cpp
@@ -3,7 +3,7 @@
 
 using namespace std;
 
-long util::gcd_l(long a, long b)
+long long util::gcd_l(long long a, long long b)
 {
 	for (;;)
 	{
@@ -17,14 +17,14 @@ long util::gcd_l(long a, long b)
 	throw exception("gcd not found");
 }
 
-long util::gcd(set<long>::iterator first, set<long>::iterator last)
+long long util::gcd(set<long long>::iterator first, set<long long>::iterator last)
 {
 	long l1 = *first;
 	if (++first == last) return l1;
 	else return gcd_l(abs(l1), abs(gcd(first, last)));
 }
 
-long util::gcd(const set<long> s)
+long long util::gcd(const set<long long> s)
 {
 	return gcd(s.begin(), s.end());
 }
