+++ euler579c/vector3d.h
@@ -10,13 +10,13 @@ using namespace std;
 class vector3d : public vertex
 {
 public:
-	unsigned long length;
-	vector3d(long X, long Y, long Z);
+	long long length;
+	vector3d(long long X, long long Y, long long Z);
 
 	bool is_orthogonal_to(const vector3d&) const;
-	unsigned long dot_product(const vector3d&) const;
+	long long dot_product(const vector3d&) const;
 	vector3d cross_product(const vector3d&) const;
-	long gcd() const;
+	long long gcd() const;
 
 };
